Fix players.birth_year CHECK constraint, and completed correspondinginclusions and updates required

+++ supabase/schema.sql
@@ -130,7 +130,7 @@ CREATE TABLE players (
   birth_year integer 
     CHECK (
       birth_year >= 1850                  -- Enforcing time (year) range
-      AND birth_year <= EXTRACT(YEAR FROM CURRENT_DATE - 15)
+      AND birth_year <= (EXTRACT(YEAR FROM CURRENT_DATE)::int - 15)
     ),
   notes text                              -- Any other information
 );
